% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_strategy.R
\name{by_group}
\alias{by_group}
\title{Define a formula whose value is group-dependent}
\usage{
by_group(.group, ...)
}
\arguments{
\item{.group}{Optional group name. If not specified
it is implicitly added.}

\item{...}{Values of the parameter named depending on the
group}
}
\value{
A vector of values.
}
\description{
Returns different values depending on the group
}
\examples{

define_parameters(
  val = 456,
  x = by_group(
    strat_1 = 1234,
    strat_2 = 9876,
    strat_3 = val * 2 + markov_cycle
  )
)
}
