% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_strategy.R
\name{by_group_hack}
\alias{by_group_hack}
\title{Hack to Automate Use of Group Name}
\usage{
by_group_hack(.dots)
}
\arguments{
\item{.dots}{A \code{lazy_dots} object.}
}
\value{
A modified \code{lazy_dots} object.
}
\description{
This function is a hack to automate the definition of the
argument \code{.group} in
\code{\link[=by_group]{by_group()}}.
}
\details{
The hack consists in replacing calls to
\code{by_group(...)} by
\code{by_group(.group = group, ...)} if
\code{.group} is not already defined.
}
\keyword{internal}
