% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_eval.R
\name{c_linear}
\alias{c_linear}
\title{Linearization of cost}
\usage{
c_linear(res_vbp, strategy)
}
\arguments{
\item{res_vbp}{The result of a VBP analysis}

\item{strategy}{Strategy}
}
\value{
Intercept and slope of linearization, and indicator whether
linearization was performed
}
\description{
Calculates the linear equation of cost as a function of price per strategy
}
