% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_eval.R
\name{ce_strategy}
\alias{ce_strategy}
\title{Cost and effectiveness per strategy}
\usage{
ce_strategy(model, strategy)
}
\arguments{
\item{model}{A heemod model}

\item{strategy}{Strategy}
}
\value{
Cost and effectivness for selected strategy
}
\description{
Returns the cost and effectiveness for selected strategy
}
