% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_eval.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Compute Expected Value of partial Perfect Information (EVPPI)}
\usage{
compute_evppi(x, evppi, max_wtp = 100000, n = 10, verbose = TRUE)
}
\arguments{
\item{x}{Result from \code{\link[=run_psa]{run_psa()}}.}

\item{evppi}{An object returned by
\code{\link[=define_evppi]{define_evppi()}}.}

\item{max_wtp}{Maximal willingness to pay.}

\item{n}{Number of WTP thresholds to estimate EVPPI (values above
10 may take significant time).}

\item{verbose}{Logical variable indicating whether to display the progress
of EVPPI calculations}
}
\value{
A \code{data.frame} with one row per WTP threshold
}
\description{
Compute EVPPI using a linear regression metamodeling approach.
}
\references{
\enumerate{
\item Jalal H, Alarid-Escudero F. A Gaussian Approximation Approach for Value of Information Analysis. Med Decis Making 2018; 38(2): 174-188.
}
}
