% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_define.R
\name{define_evppi}
\alias{define_evppi}
\alias{define_evppi_}
\title{Define an Expected Value of Partial Perfect Information (EVPPI) Analysis}
\usage{
define_evppi(...)

define_evppi_(par_name)
}
\arguments{
\item{par_name}{String of parameter name}
}
\value{
An \code{evppi} object.
}
\description{
Define the parameter for EVPPI analysis of Markov models using a
linear regression metmodeling approach
}
\details{
The parameter name should be a string
}
\references{
\enumerate{
\item Jalal H, Alarid-Escudero F. A Gaussian Approximation Approach for Value of Information Analysis.
Med Decis Making 2017;OnlineFirst.
}
}
