% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_surv_function}
\alias{define_surv_function}
\title{Define a survival distribution based on a function}
\usage{
define_surv_function(f, ...)
}
\arguments{
\item{f}{a function that generates a vector of survival probabilities}
}
\value{
a \code{surv_function} object, which can be used with \code{\link[=compute_surv]{compute_surv()}}.
}
\description{
Define a survival distribution based on a function
}
