% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_surv_lifetable}
\alias{define_surv_lifetable}
\alias{define_surv_lifetable.data.frame}
\title{Define a survival distribution based on explicit survival probabilities}
\usage{
define_surv_lifetable(x, start_age, percent_male)

\method{define_surv_lifetable}{data.frame}(x, start_age, percent_male)
}
\arguments{
\item{x}{a data frame with columns \code{age}, \code{male}, and \code{female}, where age is
a counting sequence representing each age covered by the life-table, male
represents the conditional probability of death at each age for men, and female represents
the conditional probability  of death at each age for women.}
}
\value{
a \code{surv_lifetable} object, which can be used with \code{\link[=compute_surv]{compute_surv()}}.
}
\description{
Define a survival distribution based on explicit survival probabilities
}
\examples{
 x <- data.frame(age = c(0, 1, 2, 3), male = c(0.011, 0.005, 0.003, 0.002), female = c(0.010, 0.005, 0.004, 0.002))
 define_surv_lifetable(x, 1, 0.45)
 
}
