% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_survival_cure}
\alias{define_survival_cure}
\title{Define a Parametric Mixture or Non-Mixture Cure Distribution}
\usage{
define_survival_cure(
  distribution = c("exp", "weibull", "weibullPH", "lnorm", "llogis", "gamma",
    "gompertz", "gengamma", "gengamma.orig", "genf", "genf.orig"),
  theta,
  ...,
  mixture = T
)
}
\arguments{
\item{distribution}{A parametric survival distribution.}

\item{theta}{The model cure fraction.}

\item{...}{Additional distribution parameters (see
respective distribution help pages).}

\item{mixture}{a logical determining whether a mixture
or non-mixture is being defined.}
}
\value{
A \code{surv_dist_cure} object.
}
\description{
Define a parametric cure survival model.
}
\examples{

define_survival_cure(distribution = "exp", theta = 0.34, rate = .5)
define_survival_cure(distribution = "weibull", theta = 0.5, shape = 1.5, scale = 34.43)

}
