% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_define.R
\name{define_vbp}
\alias{define_vbp}
\alias{define_vbp_}
\title{Define a Value-Based Pricing Analysis}
\usage{
define_vbp(...)

define_vbp_(par_name, low_dots, med_dots, high_dots)
}
\arguments{
\item{...}{A parameter name and min/max values
of the form \code{price, min(price), max(price)}.}

\item{par_name}{String vector of price parameter name.}

\item{low_dots, med_dots, high_dots}{Used to work around
non-standard evaluation.}
}
\value{
A \code{vbp} object.
}
\description{
Define parameter variations for a value-based
pricing analysis.
}
\examples{

define_vbp(
  p, 0, 1000
)

}
