% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{eval_surv}
\alias{eval_surv}
\alias{compute_surv_}
\alias{surv_prob_}
\alias{event_prob_}
\alias{eval_surv.default}
\alias{eval_surv.surv_function}
\alias{eval_surv.survfit}
\alias{eval_surv.flexsurvreg}
\alias{eval_surv.surv_model}
\alias{eval_surv.surv_projection}
\alias{eval_surv.surv_pooled}
\alias{eval_surv.surv_ph}
\alias{eval_surv.surv_shift}
\alias{eval_surv.surv_aft}
\alias{eval_surv.surv_po}
\alias{eval_surv.surv_add_haz}
\alias{eval_surv.surv_dist}
\alias{eval_surv.surv_dist_cure}
\alias{eval_surv.surv_dist_spline}
\alias{eval_surv.surv_table}
\alias{eval_surv.surv_lifetable}
\title{Evaluate Survival Distributions}
\usage{
eval_surv(x, time, ...)

compute_surv_(x, time, cycle_length = 1, type = c("prob", "survival"), ...)

surv_prob_(x, time, ...)

event_prob_(x, start, end, ...)

\method{eval_surv}{default}(x, time, ...)

\method{eval_surv}{surv_function}(x, time, ...)

\method{eval_surv}{survfit}(x, time, ...)

\method{eval_surv}{flexsurvreg}(x, time, ...)

\method{eval_surv}{surv_model}(x, time, ...)

\method{eval_surv}{surv_projection}(x, time, ...)

\method{eval_surv}{surv_pooled}(x, time, ...)

\method{eval_surv}{surv_ph}(x, time, ...)

\method{eval_surv}{surv_shift}(x, time, ...)

\method{eval_surv}{surv_aft}(x, time, ...)

\method{eval_surv}{surv_po}(x, time, ...)

\method{eval_surv}{surv_add_haz}(x, time, ...)

\method{eval_surv}{surv_dist}(x, time, ...)

\method{eval_surv}{surv_dist_cure}(x, time, ...)

\method{eval_surv}{surv_dist_spline}(x, time, ...)

\method{eval_surv}{surv_table}(x, time, ...)

\method{eval_surv}{surv_lifetable}(x, time, ...)
}
\arguments{
\item{x}{A survival distribution object}

\item{time}{The \code{model_time} or \code{state_time} for which
to predict.}

\item{...}{arguments passed to methods.}

\item{cycle_length}{The value of a Markov cycle in
absolute time units.}

\item{type}{Either \code{prob}, for transition probabilities,
or \code{surv}, for survival.}

\item{start}{The start time of the interval.}

\item{end}{The end time of the interval.}
}
\value{
Returns either the survival probalities or
conditional probabilities of event for each cycle.
}
\description{
Generate either survival probabilities or conditional
probabilities of event for each model cycle.
}
\details{
The results of \code{compute_surv()} are memoised for
\code{options("heRomod.memotime")} (default: 1 hour) to
increase resampling performance.
}
\keyword{internal}
