% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{event_prob}
\alias{event_prob}
\title{Evaluate Event Probabilities}
\usage{
event_prob(x, start, end, ...)
}
\arguments{
\item{x}{A survival distribution object}

\item{start}{The end time of the interval.}

\item{...}{arguments passed to methods.}
}
\value{
Returns either the survival probalities or
conditional probabilities of event for each cycle.
}
\description{
Generate event probabilities from a survival
distribution.
}
\details{
The results of \code{event_prob()} are memoised for
\code{options("heRomod.memotime")} (default: 1 hour) to
increase resampling performance.
}
