% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_eval.R
\name{p_comp}
\alias{p_comp}
\title{Price of comparator strategy}
\usage{
p_comp(e.P, e.comp, beta0.P, beta0.comp, beta1.P, beta1.comp, lambda)
}
\arguments{
\item{e.P}{Effectiveness of strategy of interest}

\item{e.comp}{Effectiveness of comparator strategy}

\item{beta0.P}{Intercept of strategy of interest}

\item{beta0.comp}{Intercept of comparator strategy}

\item{beta1.P}{Slope of strategy of interest}

\item{beta1.comp}{Slope of comparator strategy}

\item{lambda}{Vector of willingness to pay thresholds}
}
\value{
Vector of prices of comparator strategy
}
\description{
Returns the cost and effectiveness for selected strategy
}
