% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_eval.R
\name{param_in_strategy}
\alias{param_in_strategy}
\title{Parameter influences strategy}
\usage{
param_in_strategy(mod, strategy, parameter)
}
\arguments{
\item{mod}{An evaluated Markov model}

\item{strategy}{Strategy of interest}

\item{parameter}{Parameter of interest}
}
\value{
TRUE if parameter potentially influences strategy, FALSE otherwise
}
\description{
Determines if parameter influences strategy
}
