% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_print.R
\name{plot.evppi_res}
\alias{plot.evppi_res}
\title{Plot Results of Expected Value of Partial Perfect Information Analysis}
\usage{
\method{plot}{evppi_res}(x, bw = FALSE, ...)
}
\arguments{
\item{x}{Result from \code{\link[=compute_evppi]{compute_evppi()}}.}

\item{bw}{Black & white plot for publications?}

\item{...}{Additional arguments, depends on \code{type}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot of EVPPI for one or more parameters by willingness to pay threshold
}
