% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_print.R
\name{plot.vbp}
\alias{plot.vbp}
\title{Plot Sensitivity Analysis}
\usage{
\method{plot}{vbp}(x, bw = FALSE, ...)
}
\arguments{
\item{x}{A result of \code{\link[=run_vbp]{run_vbp()}}.}

\item{bw}{Black & white plot for publications?}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the results of a sensitivity analysis as a tornado
plot.
}
\details{
Plot type \code{simple} plots variations of single strategy
values, while \code{difference} plots incremental values.
}
