% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_dependencies.default}
\alias{resolve_dependencies.default}
\title{Reorder a list of expressions to resolve dependencies}
\usage{
\method{resolve_dependencies}{default}(x)
}
\description{
Reorder a list of expressions to resolve dependencies
}
