% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardise_call}
\alias{standardise_call}
\title{Hack to Work Around a Discounting Issue}
\usage{
standardise_call(call, env = parent.frame())
}
\arguments{
\item{.dots}{A state object.}
}
\value{
A modified state object.
}
\description{
This function is a hack to avoid a problem with
discounting when the argument is a constant.
}
\details{
The hack consists in replacing calls to
\code{discount(x)} by \code{discount(x * rep(1, n()))} to
ensure \code{x} is recycled to the correct length.
}
\keyword{internal}
