% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbp_print.R
\name{summary.vbp}
\alias{summary.vbp}
\title{Summarise Value-Based Pricing Results}
\usage{
\method{summary}{vbp}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=run_vbp]{run_vbp()}}.}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
A \code{summary_vbp} object.
}
\description{
Summarise Value-Based Pricing Results
}
