% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{surv_prob}
\alias{surv_prob}
\title{Evaluate Survival Probabilities}
\usage{
surv_prob(x, time, ...)
}
\arguments{
\item{x}{A survival distribution object}

\item{time}{The times for which to predict.}

\item{...}{arguments passed to methods.}
}
\value{
Returns either the survival probalities or
conditional probabilities of event for each cycle.
}
\description{
Generate survival probabilities from a survival
distribution.
}
\details{
The results of \code{surv_prob()} are memoised for
\code{options("heRomod.memotime")} (default: 1 hour) to
increase resampling performance.
}
