% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_eval.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Compute Expected Value of partial Perfect Information (EVPPI)}
\usage{
compute_evppi(x, evppi, max_wtp = 1e+05, n = 10, verbose = TRUE)
}
\arguments{
\item{x}{Result from \code{\link[=run_psa]{run_psa()}}.}

\item{evppi}{An object returned by
\code{\link[=define_evppi]{define_evppi()}}.}

\item{max_wtp}{Maximal willingness to pay.}

\item{n}{Number of WTP thresholds to estimate EVPPI (values above
10 may take significant time).}

\item{verbose}{Logical variable indicating whether to display the progress
of EVPPI calculations}
}
\value{
A \code{data.frame} with one row per WTP threshold
}
\description{
Compute EVPPI using a linear regression metamodeling approach.
}
\examples{
# example for run_evppi

mod1 <- define_strategy(
  transition = define_transition(
    .5, .5,
    .1, .9
  ),
  define_state(
    cost = cost_init + age * 5,
    ly = 1
  ),
  define_state(
    cost = cost_init + age,
    ly = 0
  )
)

mod2 <- define_strategy(
  transition = define_transition(
    p_trans, C,
    .1, .9
  ),
  define_state(
    cost = 789 * age / 10,
    ly = 1
  ),
  define_state(
    cost = 456 * age / 10,
    ly = 0
  )
  
)

res2 <- run_model(
  mod1, mod2,
  parameters = define_parameters(
    age_init = 60,
    cost_init = 1000,
    age = age_init + markov_cycle,
    p_trans = .7
  ),
  init = 1:0,
  cycles = 10,
  cost = cost,
  effect = ly
)

rsp <- define_psa(
  age_init ~ normal(60, 10),
  cost_init ~ normal(1000, 100),
  p_trans ~ binomial(.7, 100),
  correlation = matrix(c(
    1,  .4, 0,
    .4, 1,  0,
    0,  0,  1
  ), byrow = TRUE, ncol = 3)
)


# with run_model result
# Could take a while to produce 1000 samples per strategy
ndt1 <- run_psa(res2, psa = rsp, N = 1000)

# Plot EVPI
plot(ndt1, type = "evpi", max_wtp = 50000, log_scale = FALSE)

### EVPPI of one parameter
# Define evppi
def_evppi <- define_evppi(
  p_trans
)

## For one WTP
# Compute evppi on one WTP threshold value
evppi1 <- compute_evppi(x = ndt1, 
                        evppi = def_evppi, 
                        max_wtp = 20000, n = 1)

evppi1

# Plot evppi 
plot(evppi1)

# Plot evppi bw
plot(evppi1, bw = TRUE)

## For 100 WTP threshold values
evppi2 <- compute_evppi(x = ndt1, evppi = def_evppi, 
                        max_wtp = 50000, n = 100)
evppi2

# Plot evppi
plot(evppi2)

# Plot evppi bw
plot(evppi2, bw = TRUE)

### EVPPI of three parameters
## For one WTP
# Define evppi
def_evppi <- define_evppi(
  p_trans,
  cost_init,
  age_init
)
def_evppi

# Compute evppi on one WTP threshold value
evppi1 <- compute_evppi(x = ndt1, 
                        evppi = def_evppi, 
                        max_wtp = 20000, n = 1)

evppi1

plot(evppi1)

plot(evppi1, bw = TRUE)

## For 100 WTP threshold values
evppi2 <- compute_evppi(x = ndt1, evppi = def_evppi, 
                        max_wtp = 50000, n = 100)
evppi2

# Plot evppi
plot(evppi2)

# Plot evppi bw
plot(evppi2, bw = TRUE)
}
\references{
\enumerate{
\item Jalal H, Alarid-Escudero F. A Gaussian Approximation Approach for Value of Information Analysis.
Med Decis Making 2018; 38: 174–188.
}
}
