% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heRomod.R
\docType{package}
\name{heRomod}
\alias{heRomod}
\title{heRomod: Health Economic Evaluation MODelling}
\description{
\code{heRomod} is an R toolset for health economic
evaluation modelling. It aims to provide a simple and
consistent interface for Markov models specification and
comparison. Non-homogeneous Markov models (with time
varying properties) are supported.
}
